Added UI and support for multiple background audio clips
Added basic UI to the main settings menu to quickly toggle between background audio clips and adjust the volume of the current clip. There is no background audio by default until it is toggled on. On the developer's end it should be easy to add as many more audio clips as necessary, and I have found two for now.